I found a bug where running with no work directory resulted in an error:
ERROR ~ Error executing process > 'IngestPairs:QCReads:PairtoolsStats (4)'

Caused by:
  Not a valid path value: 'KO_2_1_fragtag.pairs.gz'

QCReads looks for keys including frag_pairs to
generate stats and accumulate in a multiQC report.
However it wasn't checking whether or not the value
is actually a file.

The OptionalFragtag workflow adds a frag_pairs
key to relevant sample channel items but it's just
a value storing the name of the file to create.
When this gets added, it is seen also by QCReads
as called by the previous IngestPairs workflow
and therefore, QCReads tries to pass it as the
path value to the PairtoolsStats process. This
fails because PairtoolsStats expects a path, not
a value.

To fix this, I added some code to validate that
all hashmap values associated with keys like frag_pairs are
files, not vals. However, the deeper issue remains
that an earlier step in the workflow is seeing
key:value pairs in the channel that are associated
with later steps in the workflow. This should
probably be fixed if possible.