Restructure of fragtag script
Returning to the code to figure out a bug,
the fragtag script was nearly incomprehensible.

For the restructure, I treat the program as a
set of transitions between dataframes of specific
structures:

BedpePairs: A standard 4DN-structured .pairs format
BedpeIDPairs: Same, but contains a pairID index column
BedpeIDEnds: Stores an "end" column and removes
any column name end-specific identifiers (i.e.
removes the 1 and 2 from chrom1/2, pos1/2, strand1/2)

The BedpeIDEnds contains code to tag the fragments
using the FragIndex class.

Note that the fragment tagging is currently
inaccurate. The fragment number is slightly off
from pairtools restrict and the actual start and
end is more substantially incorrect.

But the code is legible at least!