Refactored JoinProcessResults to transpack
The transpack "operator" combines two sub-"operators",
transact and pack.

Transact is an adapter operator that extracts a
subset of hashmap values, calls a process with them,
and rebuilds a new hashmap with the outputs.

Pack iteratively joins a list of channels linked
by key in order from "smallest" to "largest".

Transpack uses transact followed by pack for convenience.