Replaced MakeResourceFile with source in extraops.
source internalizes the processing formerly done
by individual staging workflows like
MakeMissingChromsizes. It separates out channel
items where the resource isn't needed according
to a supplied rule (which defaults to assuming
every hashmap needs the resource), then assigns
default names to resources where it's not given
by the user according to a second rule.

The operator also takes arguments for transpack
in order to appropriately assign the created
resource to those that were missing it. Finally
it returns the concatenated items as a single
channel.