Improved emptyOnLastStep, humid
The humid parameter now works again and is accessed
from params directly. It sets n_reads to params.general.humidDefault
if true, to itself if an Integer, or doesn't add n_reads otherwise.

Also, on the last step, if viewLastStep is given, a submap on a space-separated
list of keys will be shown for each sample in the sample channel for the output
of the last step.