Hich features, optimizations, unit tests
- Added param-based tests with Hypothesis
- Optimized PairsSegment creation and PairsClassifier .classify
  to substantially decrease runtime.
- Added hich functions to compute and aggregate stats
- Added functionality to downsample pairs files

The idea now is that for downsampling, you would compute an initial
stats file for each .pairs file. Then you'd aggregate them as desired,
such as aggregating each condition's techreps to the mean distribution
and then to the minimum size. Alternatively, downsample to the mean
subgroup distribution for each condition, then to the minimum size across
all files in a second round of aggregation. Ultimately you end up with
both a start and target distribution for each pairs file, which are
used for the actual downsampling.