Add hich reshape - edit .pairs columns
The immediate motivation for hich reshape is to
permit users to extract a per-cell barcode from
a larger value in a column and move the barcode
to a cellID column.

Hich reshape generalizes to permit several ways
of reshaping a .pairs file, including using
Python's parse() syntax or a regex to extract
text from one column and move it to another,
replacing all values in a column with a placeholder,
selecting and reordering a subset of columns,
and dropping columns.

It is built with my new small smart_open_with_pbgzip
library which replaces the default .gz/.gzip writer
with pbgzip, which will help integrate this with
Pairtools.