Implemented single-cell labeling + deduplication.
On parsing and ingesting pairs, the user can specify a value for cellBarcodeField for the sample,
along with a regex or Python parse library pattern to extract it. The barcode will be extracted
from the pairs column whose name matches cellBarcodeField using the pattern, and placed in
the cellID column. The isSingleCell flag is set to true for the sample.

During deduplication, the --extra-col-pair cellID cellID flag only identifies reads as duplicates
if they have a matching cellID. We use --backend cython because the other backends are extremely
unperformant (i.e. 3.3 minutes/million reads), due to iterating sequentially through each cellID
for building the KDTree and performing the radius search. Using cython makes deduplication non-transitive,
but I don't see an obvious argument why one is better than the other from a biological standpoint.
I also increase the batch size substantially in order to deal with the fact that each batch will be
divided up between a large number of cells.