transact now takes groupTuple outputs
Individual process instances may need to take or emit multiple samples.
Transact has been altered to facilitate this:

Process structure for processes taking and/or emitting multiple samples:

    Processes requiring multiple sample inputs should be structured
    to accept, for each input argument, a list of values. The values at a
    particular index should be treated as coming from the same sample.

    Processes emitting multiple output samples should be structured
    similarly, with each output argument being a list where the value at
    a particular index all belong to the same sample. This requires that
    every process output is a list, and all these lists have uniform lengths.

To accomplish this, transact was altered to detect inputs formatted
like the output from groupTuple() with default parameters, called as follows:
samples | map{samples.groupKey, samples} | groupTuple

This yields the result:
    [group, [HashMaps in group]]

If a channel item passed to transact is in this format, then the values
associated with each input key in [HashMaps in group] will be coalesced
into a list in the order the HashMaps appeared in the list. These lists
will be passed as the inputs to the process.

If the option 'transpose' is passed, then rather than treating each output
value as a sample, the outputs will be transposed (which requires equal
numbers of outputs for each argument), yielding a number of output tuples
equal to the length of the output arguments. These will then all be built
into output HashMaps.