Redesigned transposeHashMapList to be more general
transposeHashMapList takes a channel containing lists of hashmaps,
as after a groupTuple on a sample channel, and returns a single
hashmap in "column order". Specifically, for a hashmap H_i at index i
in the list of hashmaps L, it constructs a single output hashmap O
where for each key k_j at index j in 'keys', the value is a list v_j
where v_j_i = H_i[k]. The output hashmap has a list of values at each key k
where at a given index in the values list, the item is the value at k forthe
corresponding input hashmap in the list.