buildCmd testing and devdocs updates
I focused on testing buildCmd because it is the
most complex code in the workflow, and on
code governance documentation.

- Built tests/scripts/CommandChecker.groovy to
facilitate unit testing buildCmd methods.
- Wrote unit tests for Align and ParseToPairs
buildCmd functions.

FUTURE DIRECTIONS
A major dev pain point is writing, documenting,
validating, and testing buildCmd methods,
expected also to create pain for users. Therefore
at end of week, I decided to implement a central
YAML schema for sample attributes. This will
declare sample attributes, how to validate them,
which processes require them, and user-facing
documentation that can be automatically extracted.
Samples will be validated automatically based on
this schema using a generic validator when
attributes are extracted from samples to get
for process inputs. This will be the primary
focus of the next sprint.